fix: rename file structure
