+++ src/data-structures/queue.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "linked-list.h"
+#include "queue.h"
 
 static Node new_node(Position p) {
   Node node = NULL;
